==> Crypto_factor_ML_Pricing_Model/01_data_prep/02_get_raw_individualcoins_data.py <==
# This file fetches daily historical data for the top 20 cryptocurrencies for the year 2025

import os
from dotenv import load_dotenv
from requests import Session
import pandas as pd
import time

# ==================== LOAD API KEY ====================
load_dotenv()
API_KEY = os.getenv('CMC_API_KEY')
if not API_KEY:
    raise ValueError("API key not found in .env!")

# ==================== YOUR ORIGINAL DATA (unchanged) ====================
data = {'data': [{'value': 193.67, 'update_time': '2025-11-28T00:00:00Z', 'constituents': [{'id': 1, 'priceUsd': 91327.67331120375, 'units': 0.00144496809438, 'name': 'Bitcoin', 'symbol': 'BTC', 'url': 'https://coinmarketcap.com/currencies/bitcoin', 'weight': 68.18}, {'id': 1027, 'priceUsd': 3015.513478626419, 'units': 0.00874637489015, 'name': 'Ethereum', 'symbol': 'ETH', 'url': 'https://coinmarketcap.com/currencies/ethereum', 'weight': 13.62}, {'id': 52, 'priceUsd': 2.2006828870047968, 'units': 4.35260345599326, 'name': 'XRP', 'symbol': 'XRP', 'url': 'https://coinmarketcap.com/currencies/xrp', 'weight': 4.97}, {'id': 1839, 'priceUsd': 895.5397705904559, 'units': 0.00998481881508, 'name': 'BNB', 'symbol': 'BNB', 'url': 'https://coinmarketcap.com/currencies/bnb', 'weight': 4.61}, {'id': 5426, 'priceUsd': 140.84448696481846, 'units': 0.03980355172901, 'name': 'Solana', 'symbol': 'SOL', 'url': 'https://coinmarketcap.com/currencies/solana', 'weight': 2.95}, {'id': 1958, 'priceUsd': 0.280232817003908, 'units': 6.88141328811557, 'name': 'TRON', 'symbol': 'TRX', 'url': 'https://coinmarketcap.com/currencies/tron', 'weight': 0.99}, {'id': 74, 'priceUsd': 0.15267228390994253, 'units': 10.9350291661837, 'name': 'Dogecoin', 'symbol': 'DOGE', 'url': 'https://coinmarketcap.com/currencies/dogecoin', 'weight': 0.87}, {'id': 2010, 'priceUsd': 0.43356547104766924, 'units': 2.61578403123211, 'name': 'Cardano', 'symbol': 'ADA', 'url': 'https://coinmarketcap.com/currencies/cardano', 'weight': 0.58}, {'id': 32196, 'priceUsd': 34.71967687406683, 'units': 0.02416802644623, 'name': 'Hyperliquid', 'symbol': 'HYPE', 'url': 'https://coinmarketcap.com/currencies/hyperliquid', 'weight': 0.44}, {'id': 1831, 'priceUsd': 532.4742054082026, 'units': 0.00144220734128, 'name': 'Bitcoin Cash', 'symbol': 'BCH', 'url': 'https://coinmarketcap.com/currencies/bitcoin-cash', 'weight': 0.4}, {'id': 1975, 'priceUsd': 13.34703599043677, 'units': 0.05029290693256, 'name': 'Chainlink', 'symbol': 'LINK', 'url': 'https://coinmarketcap.com/currencies/chainlink', 'weight': 0.35}, {'id': 512, 'priceUsd': 0.25560023417673217, 'units': 2.30506117896973, 'name': 'Stellar', 'symbol': 'XLM', 'url': 'https://coinmarketcap.com/currencies/stellar', 'weight': 0.31}, {'id': 1437, 'priceUsd': 489.5825288444158, 'units': 0.00115707740888, 'name': 'Zcash', 'symbol': 'ZEC', 'url': 'https://coinmarketcap.com/currencies/zcash', 'weight': 0.3}, {'id': 2, 'priceUsd': 86.62461828657932, 'units': 0.00564312304791, 'name': 'Litecoin', 'symbol': 'LTC', 'url': 'https://coinmarketcap.com/currencies/litecoin', 'weight': 0.25}, {'id': 5805, 'priceUsd': 15.023447405994505, 'units': 0.03091287921223, 'name': 'Avalanche', 'symbol': 'AVAX', 'url': 'https://coinmarketcap.com/currencies/avalanche', 'weight': 0.24}, {'id': 4642, 'priceUsd': 0.1471013414449339, 'units': 3.02894967303303, 'name': 'Hedera', 'symbol': 'HBAR', 'url': 'https://coinmarketcap.com/currencies/hedera', 'weight': 0.23}, {'id': 20947, 'priceUsd': 1.5431376595728838, 'units': 0.26761799308817, 'name': 'Sui', 'symbol': 'SUI', 'url': 'https://coinmarketcap.com/currencies/sui', 'weight': 0.21}, {'id': 5994, 'priceUsd': 8.763223521649563e-06, 'units': 42076.08647614162, 'name': 'Shiba Inu', 'symbol': 'SHIB', 'url': 'https://coinmarketcap.com/currencies/shiba-inu', 'weight': 0.19}, {'id': 3635, 'priceUsd': 0.11137422045249087, 'units': 2.57597781888021, 'name': 'Cronos', 'symbol': 'CRO', 'url': 'https://coinmarketcap.com/currencies/cronos', 'weight': 0.16}, {'id': 11419, 'priceUsd': 1.634887768921494, 'units': 0.18483543326042, 'name': 'Toncoin', 'symbol': 'TON', 'url': 'https://coinmarketcap.com/currencies/toncoin', 'weight': 0.15}]}, {'value': 192.78, 'update_time': '2025-11-29T00:00:00Z', 'constituents': [{'id': 1, 'priceUsd': 90915.70640237529, 'units': 0.00144496809438, 'name': 'Bitcoin', 'symbol': 'BTC', 'url': 'https://coinmarketcap.com/currencies/bitcoin', 'weight': 68.19}, {'id': 1027, 'priceUsd': 3031.905573524726, 'units': 0.00874637489015, 'name': 'Ethereum', 'symbol': 'ETH', 'url': 'https://coinmarketcap.com/currencies/ethereum', 'weight': 13.75}, {'id': 52, 'priceUsd': 2.180491549686994, 'units': 4.35260345599326, 'name': 'XRP', 'symbol': 'XRP', 'url': 'https://coinmarketcap.com/currencies/xrp', 'weight': 4.94}, {'id': 1839, 'priceUsd': 887.6084735627275, 'units': 0.00998481881508, 'name': 'BNB', 'symbol': 'BNB', 'url': 'https://coinmarketcap.com/currencies/bnb', 'weight': 4.6}, {'id': 5426, 'priceUsd': 137.39210164459337, 'units': 0.03980355172901, 'name': 'Solana', 'symbol': 'SOL', 'url': 'https://coinmarketcap.com/currencies/solana', 'weight': 2.89}, {'id': 1958, 'priceUsd': 0.2810957009971356, 'units': 6.88141328811557, 'name': 'TRON', 'symbol': 'TRX', 'url': 'https://coinmarketcap.com/currencies/tron', 'weight': 1.0}, {'id': 74, 'priceUsd': 0.150275771455807, 'units': 10.9350291661837, 'name': 'Dogecoin', 'symbol': 'DOGE', 'url': 'https://coinmarketcap.com/currencies/dogecoin', 'weight': 0.86}, {'id': 2010, 'priceUsd': 0.4202165559537653, 'units': 2.61578403123211, 'name': 'Cardano', 'symbol': 'ADA', 'url': 'https://coinmarketcap.com/currencies/cardano', 'weight': 0.57}, {'id': 32196, 'priceUsd': 34.91909987355754, 'units': 0.02416802644623, 'name': 'Hyperliquid', 'symbol': 'HYPE', 'url': 'https://coinmarketcap.com/currencies/hyperliquid', 'weight': 0.44}, {'id': 1831, 'priceUsd': 549.3801270316709, 'units': 0.00144220734128, 'name': 'Bitcoin Cash', 'symbol': 'BCH', 'url': 'https://coinmarketcap.com/currencies/bitcoin-cash', 'weight': 0.41}, {'id': 1975, 'priceUsd': 13.13843197242166, 'units': 0.05029290693256, 'name': 'Chainlink', 'symbol': 'LINK', 'url': 'https://coinmarketcap.com/currencies/chainlink', 'weight': 0.34}, {'id': 512, 'priceUsd': 0.25346007676150434, 'units': 2.30506117896973, 'name': 'Stellar', 'symbol': 'XLM', 'url': 'https://coinmarketcap.com/currencies/stellar', 'weight': 0.31}, {'id': 1437, 'priceUsd': 452.16009292752597, 'units': 0.00115707740888, 'name': 'Zcash', 'symbol': 'ZEC', 'url': 'https://coinmarketcap.com/currencies/zcash', 'weight': 0.28}, {'id': 2, 'priceUsd': 84.30945029944162, 'units': 0.00564312304791, 'name': 'Litecoin', 'symbol': 'LTC', 'url': 'https://coinmarketcap.com/currencies/litecoin', 'weight': 0.24}, {'id': 5805, 'priceUsd': 14.852629850519476, 'units': 0.03091287921223, 'name': 'Avalanche', 'symbol': 'AVAX', 'url': 'https://coinmarketcap.com/currencies/avalanche', 'weight': 0.24}, {'id': 4642, 'priceUsd': 0.1453753809433614, 'units': 3.02894967303303, 'name': 'Hedera', 'symbol': 'HBAR', 'url': 'https://coinmarketcap.com/currencies/hedera', 'weight': 0.23}, {'id': 20947, 'priceUsd': 1.5221852622926229, 'units': 0.26761799308817, 'name': 'Sui', 'symbol': 'SUI', 'url': 'https://coinmarketcap.com/currencies/sui', 'weight': 0.21}, {'id': 5994, 'priceUsd': 8.651617004091901e-06, 'units': 42076.08647614162, 'name': 'Shiba Inu', 'symbol': 'SHIB', 'url': 'https://coinmarketcap.com/currencies/shiba-inu', 'weight': 0.19}, {'id': 3635, 'priceUsd': 0.10879437118668929, 'units': 2.57597781888021, 'name': 'Cronos', 'symbol': 'CRO', 'url': 'https://coinmarketcap.com/currencies/cronos', 'weight': 0.15}, {'id': 11419, 'priceUsd': 1.5860586560012409, 'units': 0.18483543326042, 'name': 'Toncoin', 'symbol': 'TON', 'url': 'https://coinmarketcap.com/currencies/toncoin', 'weight': 0.15}]}, {'value': 192.09, 'update_time': '2025-11-30T00:00:00Z', 'constituents': [{'id': 1, 'priceUsd': 90846.58191856531, 'units': 0.00144496809438, 'name': 'Bitcoin', 'symbol': 'BTC', 'url': 'https://coinmarketcap.com/currencies/bitcoin', 'weight': 68.38}, {'id': 1027, 'priceUsd': 2990.7578238103056, 'units': 0.00874637489015, 'name': 'Ethereum', 'symbol': 'ETH', 'url': 'https://coinmarketcap.com/currencies/ethereum', 'weight': 13.62}, {'id': 52, 'priceUsd': 2.20252819873626, 'units': 4.35260345599326, 'name': 'XRP', 'symbol': 'XRP', 'url': 'https://coinmarketcap.com/currencies/xrp', 'weight': 5.01}, {'id': 1839, 'priceUsd': 873.4606187921912, 'units': 0.00998481881508, 'name': 'BNB', 'symbol': 'BNB', 'url': 'https://coinmarketcap.com/currencies/bnb', 'weight': 4.54}, {'id': 5426, 'priceUsd': 136.02355109926054, 'units': 0.03980355172901, 'name': 'Solana', 'symbol': 'SOL', 'url': 'https://coinmarketcap.com/currencies/solana', 'weight': 2.87}, {'id': 1958, 'priceUsd': 0.2810417009929901, 'units': 6.88141328811557, 'name': 'TRON', 'symbol': 'TRX', 'url': 'https://coinmarketcap.com/currencies/tron', 'weight': 1.0}, {'id': 74, 'priceUsd': 0.14850236925236435, 'units': 10.9350291661837, 'name': 'Dogecoin', 'symbol': 'DOGE', 'url': 'https://coinmarketcap.com/currencies/dogecoin', 'weight': 0.85}, {'id': 2010, 'priceUsd': 0.41511138004817255, 'units': 2.61578403123211, 'name': 'Cardano', 'symbol': 'ADA', 'url': 'https://coinmarketcap.com/currencies/cardano', 'weight': 0.56}, {'id': 32196, 'priceUsd': 34.373612334136, 'units': 0.02416802644623, 'name': 'Hyperliquid', 'symbol': 'HYPE', 'url': 'https://coinmarketcap.com/currencies/hyperliquid', 'weight': 0.44}, {'id': 1831, 'priceUsd': 521.6309674337625, 'units': 0.00144220734128, 'name': 'Bitcoin Cash', 'symbol': 'BCH', 'url': 'https://coinmarketcap.com/currencies/bitcoin-cash', 'weight': 0.39}, {'id': 1975, 'priceUsd': 12.996302896676028, 'units': 0.05029290693256, 'name': 'Chainlink', 'symbol': 'LINK', 'url': 'https://coinmarketcap.com/currencies/chainlink', 'weight': 0.34}, {'id': 512, 'priceUsd': 0.2541521059153869, 'units': 2.30506117896973, 'name': 'Stellar', 'symbol': 'XLM', 'url': 'https://coinmarketcap.com/currencies/stellar', 'weight': 0.31}, {'id': 1437, 'priceUsd': 460.0134108550169, 'units': 0.00115707740888, 'name': 'Zcash', 'symbol': 'ZEC', 'url': 'https://coinmarketcap.com/currencies/zcash', 'weight': 0.28}, {'id': 2, 'priceUsd': 84.02829061500208, 'units': 0.00564312304791, 'name': 'Litecoin', 'symbol': 'LTC', 'url': 'https://coinmarketcap.com/currencies/litecoin', 'weight': 0.24}, {'id': 4642, 'priceUsd': 0.14329573636694404, 'units': 3.02894967303303, 'name': 'Hedera', 'symbol': 'HBAR', 'url': 'https://coinmarketcap.com/currencies/hedera', 'weight': 0.23}, {'id': 5805, 'priceUsd': 14.253230960233962, 'units': 0.03091287921223, 'name': 'Avalanche', 'symbol': 'AVAX', 'url': 'https://coinmarketcap.com/currencies/avalanche', 'weight': 0.23}, {'id': 20947, 'priceUsd': 1.4995619052125402, 'units': 0.26761799308817, 'name': 'Sui', 'symbol': 'SUI', 'url': 'https://coinmarketcap.com/currencies/sui', 'weight': 0.21}, {'id': 5994, 'priceUsd': 8.535152960141978e-06, 'units': 42076.08647614162, 'name': 'Shiba Inu', 'symbol': 'SHIB', 'url': 'https://coinmarketcap.com/currencies/shiba-inu', 'weight': 0.19}, {'id': 3635, 'priceUsd': 0.10716651208386754, 'units': 2.57597781888021, 'name': 'Cronos', 'symbol': 'CRO', 'url': 'https://coinmarketcap.com/currencies/cronos', 'weight': 0.15}, {'id': 11419, 'priceUsd': 1.5792229558333037, 'units': 0.18483543326042, 'name': 'Toncoin', 'symbol': 'TON', 'url': 'https://coinmarketcap.com/currencies/toncoin', 'weight': 0.15}]}, {'value': 191.03, 'update_time': '2025-12-01T00:00:00Z', 'constituents': [{'id': 1, 'priceUsd': 90370.6401799052, 'units': 0.00144567770461, 'name': 'Bitcoin', 'symbol': 'BTC', 'url': 'https://coinmarketcap.com/currencies/bitcoin', 'weight': 68.4}, {'id': 1027, 'priceUsd': 2993.8652220333774, 'units': 0.00874139590133, 'name': 'Ethereum', 'symbol': 'ETH', 'url': 'https://coinmarketcap.com/currencies/ethereum', 'weight': 13.7}, {'id': 52, 'priceUsd': 2.1584732925869625, 'units': 4.36725263374646, 'name': 'XRP', 'symbol': 'XRP', 'url': 'https://coinmarketcap.com/currencies/xrp', 'weight': 4.94}, {'id': 1839, 'priceUsd': 876.3520643507846, 'units': 0.0099776971745, 'name': 'BNB', 'symbol': 'BNB', 'url': 'https://coinmarketcap.com/currencies/bnb', 'weight': 4.58}, {'id': 5426, 'priceUsd': 133.57930459299814, 'units': 0.0404983768915, 'name': 'Solana', 'symbol': 'SOL', 'url': 'https://coinmarketcap.com/currencies/solana', 'weight': 2.83}, {'id': 1958, 'priceUsd': 0.28155965547431067, 'units': 6.85487347211812, 'name': 'TRON', 'symbol': 'TRX', 'url': 'https://coinmarketcap.com/currencies/tron', 'weight': 1.01}, {'id': 74, 'priceUsd': 0.14629600181540875, 'units': 10.94773094857122, 'name': 'Dogecoin', 'symbol': 'DOGE', 'url': 'https://coinmarketcap.com/currencies/dogecoin', 'weight': 0.84}, {'id': 2010, 'priceUsd': 0.4151059491267602, 'units': 2.62020667679633, 'name': 'Cardano', 'symbol': 'ADA', 'url': 'https://coinmarketcap.com/currencies/cardano', 'weight': 0.57}, {'id': 32196, 'priceUsd': 31.941701722823918, 'units': 0.02454451359574, 'name': 'Hyperliquid', 'symbol': 'HYPE', 'url': 'https://coinmarketcap.com/currencies/hyperliquid', 'weight': 0.41}, {'id': 1831, 'priceUsd': 541.2709801846033, 'units': 0.00144641278584, 'name': 'Bitcoin Cash', 'symbol': 'BCH', 'url': 'https://coinmarketcap.com/currencies/bitcoin-cash', 'weight': 0.41}, {'id': 1975, 'priceUsd': 12.962140647972609, 'units': 0.05007829769489, 'name': 'Chainlink', 'symbol': 'LINK', 'url': 'https://coinmarketcap.com/currencies/chainlink', 'weight': 0.34}, {'id': 512, 'priceUsd': 0.24825134285127143, 'units': 2.30532828572257, 'name': 'Stellar', 'symbol': 'XLM', 'url': 'https://coinmarketcap.com/currencies/stellar', 'weight': 0.3}, {'id': 1437, 'priceUsd': 428.2159678081541, 'units': 0.00120523376231, 'name': 'Zcash', 'symbol': 'ZEC', 'url': 'https://coinmarketcap.com/currencies/zcash', 'weight': 0.27}, {'id': 2, 'priceUsd': 82.45161326294179, 'units': 0.00555554697988, 'name': 'Litecoin', 'symbol': 'LTC', 'url': 'https://coinmarketcap.com/currencies/litecoin', 'weight': 0.24}, {'id': 4642, 'priceUsd': 0.14138771373251546, 'units': 3.10166877683577, 'name': 'Hedera', 'symbol': 'HBAR', 'url': 'https://coinmarketcap.com/currencies/hedera', 'weight': 0.23}, {'id': 5805, 'priceUsd': 13.73382795338518, 'units': 0.03047005593859, 'name': 'Avalanche', 'symbol': 'AVAX', 'url': 'https://coinmarketcap.com/currencies/avalanche', 'weight': 0.22}, {'id': 20947, 'priceUsd': 1.4988771665623424, 'units': 0.26727496070235, 'name': 'Sui', 'symbol': 'SUI', 'url': 'https://coinmarketcap.com/currencies/sui', 'weight': 0.21}, {'id': 5994, 'priceUsd': 8.3730593340024e-06, 'units': 43303.50318026012, 'name': 'Shiba Inu', 'symbol': 'SHIB', 'url': 'https://coinmarketcap.com/currencies/shiba-inu', 'weight': 0.19}, {'id': 33251, 'priceUsd': 0.15946138150178407, 'units': 1.91949215277035, 'name': 'World Liberty Financial', 'symbol': 'WLFI', 'url': 'https://coinmarketcap.com/currencies/world-liberty-financial-wlfi', 'weight': 0.16}, {'id': 11419, 'priceUsd': 1.5941742066471434, 'units': 0.17974319493081, 'name': 'Toncoin', 'symbol': 'TON', 'url': 'https://coinmarketcap.com/currencies/toncoin', 'weight': 0.15}]}, {'value': 181.54, 'update_time': '2025-12-02T00:00:00Z', 'constituents': [{'id': 1, 'priceUsd': 86328.2501646382, 'units': 0.00144567770461, 'name': 'Bitcoin', 'symbol': 'BTC', 'url': 'https://coinmarketcap.com/currencies/bitcoin', 'weight': 68.75}, {'id': 1027, 'priceUsd': 2799.445316255611, 'units': 0.00874139590133, 'name': 'Ethereum', 'symbol': 'ETH', 'url': 'https://coinmarketcap.com/currencies/ethereum', 'weight': 13.48}, {'id': 52, 'priceUsd': 2.0302907976593008, 'units': 4.36725263374646, 'name': 'XRP', 'symbol': 'XRP', 'url': 'https://coinmarketcap.com/currencies/xrp', 'weight': 4.89}, {'id': 1839, 'priceUsd': 827.1089967864224, 'units': 0.0099776971745, 'name': 'BNB', 'symbol': 'BNB', 'url': 'https://coinmarketcap.com/currencies/bnb', 'weight': 4.55}, {'id': 5426, 'priceUsd': 126.68474290285394, 'units': 0.0404983768915, 'name': 'Solana', 'symbol': 'SOL', 'url': 'https://coinmarketcap.com/currencies/solana', 'weight': 2.83}, {'id': 1958, 'priceUsd': 0.2769127723919491, 'units': 6.85487347211812, 'name': 'TRON', 'symbol': 'TRX', 'url': 'https://coinmarketcap.com/currencies/tron', 'weight': 1.05}, {'id': 74, 'priceUsd': 0.13552864812324117, 'units': 10.94773094857122, 'name': 'Dogecoin', 'symbol': 'DOGE', 'url': 'https://coinmarketcap.com/currencies/dogecoin', 'weight': 0.82}, {'id': 2010, 'priceUsd': 0.38571785117190427, 'units': 2.62020667679633, 'name': 'Cardano', 'symbol': 'ADA', 'url': 'https://coinmarketcap.com/currencies/cardano', 'weight': 0.55}, {'id': 32196, 'priceUsd': 31.019446400029505, 'units': 0.02454451359574, 'name': 'Hyperliquid', 'symbol': 'HYPE', 'url': 'https://coinmarketcap.com/currencies/hyperliquid', 'weight': 0.42}, {'id': 1831, 'priceUsd': 523.129463755823, 'units': 0.00144641278584, 'name': 'Bitcoin Cash', 'symbol': 'BCH', 'url': 'https://coinmarketcap.com/currencies/bitcoin-cash', 'weight': 0.42}, {'id': 1975, 'priceUsd': 12.077992707985771, 'units': 0.05007829769489, 'name': 'Chainlink', 'symbol': 'LINK', 'url': 'https://coinmarketcap.com/currencies/chainlink', 'weight': 0.34}, {'id': 512, 'priceUsd': 0.23389483332591324, 'units': 2.30532828572257, 'name': 'Stellar', 'symbol': 'XLM', 'url': 'https://coinmarketcap.com/currencies/stellar', 'weight': 0.3}, {'id': 2, 'priceUsd': 77.57562481521568, 'units': 0.00555554697988, 'name': 'Litecoin', 'symbol': 'LTC', 'url': 'https://coinmarketcap.com/currencies/litecoin', 'weight': 0.24}, {'id': 4642, 'priceUsd': 0.13327997028762217, 'units': 3.10166877683577, 'name': 'Hedera', 'symbol': 'HBAR', 'url': 'https://coinmarketcap.com/currencies/hedera', 'weight': 0.23}, {'id': 1437, 'priceUsd': 343.82094671284455, 'units': 0.00120523376231, 'name': 'Zcash', 'symbol': 'ZEC', 'url': 'https://coinmarketcap.com/currencies/zcash', 'weight': 0.23}, {'id': 5805, 'priceUsd': 12.763363342908619, 'units': 0.03047005593859, 'name': 'Avalanche', 'symbol': 'AVAX', 'url': 'https://coinmarketcap.com/currencies/avalanche', 'weight': 0.22}, {'id': 20947, 'priceUsd': 1.3454015830299444, 'units': 0.26727496070235, 'name': 'Sui', 'symbol': 'SUI', 'url': 'https://coinmarketcap.com/currencies/sui', 'weight': 0.2}, {'id': 5994, 'priceUsd': 7.972410318593361e-06, 'units': 43303.50318026012, 'name': 'Shiba Inu', 'symbol': 'SHIB', 'url': 'https://coinmarketcap.com/currencies/shiba-inu', 'weight': 0.19}, {'id': 33251, 'priceUsd': 0.15343677811301018, 'units': 1.91949215277035, 'name': 'World Liberty Financial', 'symbol': 'WLFI', 'url': 'https://coinmarketcap.com/currencies/world-liberty-financial-wlfi', 'weight': 0.16}, {'id': 11419, 'priceUsd': 1.5034975396805481, 'units': 0.17974319493081, 'name': 'Toncoin', 'symbol': 'TON', 'url': 'https://coinmarketcap.com/currencies/toncoin', 'weight': 0.15}]}], 'status': {'timestamp': '2025-12-02T05:04:19.961Z', 'error_code': '0', 'error_message': '', 'elapsed': 100, 'credit_count': 1}}

# Extract all unique coins + nice name/symbol mapping
all_coins = [coin for day in data['data'] for coin in day['constituents']]
df_coins = pd.DataFrame(all_coins)
unique_ids = sorted(df_coins['id'].unique())
id_to_symbol = dict(zip(df_coins['id'], df_coins['symbol']))
id_to_name   = dict(zip(df_coins['id'], df_coins['name']))

print(f"Found {len(unique_ids)} unique coins → fetching daily history 2025...")

# ==================== API CALL (ONE CALL FOR ALL COINS) ====================
session = Session()
session.headers.update({
    'Accepts': 'application/json',
    'X-CMC_PRO_API_KEY': API_KEY,
})

url = 'https://pro-api.coinmarketcap.com/v2/cryptocurrency/quotes/historical'

params = {
    'id': ','.join(map(str, unique_ids)),           # ← ALL coins at once!
    'time_start': '2025-01-01T00:00:00Z',
    'time_end'  : '2025-12-02T00:00:00Z',           # today
    'interval'  : 'daily',
    'convert'   : 'USD'
}

response = session.get(url, params=params)
response.raise_for_status()
raw = response.json()

# ==================== BUILD ONE BIG DATAFRAME ====================
records = []

for coin_id_str, coin_info in raw['data'].items():
    coin_id = int(coin_id_str)
    symbol = id_to_symbol.get(coin_id, "UNKNOWN")
    name   = id_to_name.get(coin_id, "Unknown Coin")

    for q in coin_info['quotes']:
        usd = q['quote']['USD']
        records.append({
            'date'              : q['timestamp'][:10],           # YYYY-MM-DD
            'id'                : coin_id,
            'symbol'            : symbol,
            'name'              : name,
            'price_usd'         : round(usd['price'], 10),
            'volume_24h'        : usd.get('volume_24h'),
            'market_cap'        : usd.get('market_cap'),
            'percent_change_24h': usd.get('percent_change_24h'),
            'percent_change_7d' : usd.get('percent_change_7d'),
        })

df_master = pd.DataFrame(records)
df_master['date'] = pd.to_datetime(df_master['date'])
df_master = df_master.sort_values(['symbol', 'date']).reset_index(drop=True)

# ==================== SAVE BOTH FORMATS ====================

# 1. ONE BIG CSV (recommended for analysis)
master_file = '/users/a1808/Desktop/IEDA 3330/Project/Crypto_factor_ML_Pricing_Model/data/crypto_index_2025_daily_FULL.csv'
df_master.to_csv(master_file, index=False)
print(f"MASTER CSV saved → {master_file}   ({len(df_master):,} rows)")

# 2. ONE CSV PER COIN (nice folder)
folder = '/users/a1808/Desktop/IEDA 3330/Project/Crypto_factor_ML_Pricing_Model/data/daily_price_per_coin'
os.makedirs(folder, exist_ok=True)

for symbol in df_master['symbol'].unique():
    coin_df = df_master[df_master['symbol'] == symbol].copy()
    # Clean filename (no weird characters)
    safe_symbol = "".join(c if c.isalnum() else "_" for c in symbol)
    filename = f"{folder}/{safe_symbol}_2025_daily.csv"
    coin_df.to_csv(filename, index=False)

print(f"Individual files saved → {folder}/   ({len(df_master['symbol'].unique())} coins)")

# ==================== QUICK PREVIEW ====================
print("\nFirst 5 rows of master file:")
print(df_master[['date', 'symbol', 'price_usd', 'percent_change_24h']].head())

print("\nDone! You now have BOTH formats")
